+++ src/js/utils.js
@@ -1,4 +1,4 @@
 export const isBoolean = value => typeof value === "boolean";
 export const isString = value => typeof value === "string";
-export const isObject = value => typeof value === "object";
-export const between = (min, max, value) => value > min && value <= max;
+export const isObject = value => value === Object(value);
+export const isBetween = (value, min, max) => value > min && value <= max;