+++ tests/unit/ToastContainer.spec.js
@@ -5,20 +5,20 @@ import { mount, createLocalVue } from "@vue/test-utils";
 
 Vue.use(VueToastify);
 const vm = new Vue({
-  template: "<body></body>"
+    template: "<body></body>"
 });
 
 describe("VueToastify.js", () => {
-  it("updates settings", () => {
-    // Arrange
-    const wrapper = mount(ToastContainer, { vm });
+    it("updates settings", () => {
+        // Arrange
+        const wrapper = mount(ToastContainer, { vm });
 
-    // Act
-    vm.$vToastify.setSettings({ singular: true });
-    console.log(wrapper.vm.singular);
+        // Act
+        vm.$vToastify.setSettings({ singular: true });
+        console.log(wrapper.vm.singular);
 
-    // Assert
-    // expect(wrapper.singular).toBe(true);
-    expect(vm.$vToastify.getSettings().singular).toBe(true);
-  });
+        // Assert
+        // expect(wrapper.singular).toBe(true);
+        expect(vm.$vToastify.getSettings().singular).toBe(true);
+    });
 });